requirements.txt
Django==5.0.7
django-contentor-video-processor==0.8.7
python-dotenv==1.1.0
django-storages[s3]==1.14.4
requests==2.32.3
boto3==1.34.149
botocore==1.34.149